+++ repository/user_repository.go
@@ -4,7 +4,6 @@ import (
 	"database/sql"
 	"fmt"
 	"log"
-	"time"
 
 	"jmrashed/apps/userApp/model"
 )